leet-vm/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "leet-vm"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "protobuf>=4.25",
    "numpy>=1.26",
    "redis>=5.0",
    "aiohttp>=3.9",
    "pydantic>=2.5",
    "typing-extensions>=4.9",
]

[project.optional-dependencies]
core = ["leet-core"]   # PyO3 wheel — install separately
dev  = ["pytest>=7", "pytest-asyncio>=0.23"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
